superfile: add SuperfileReader::id_lookup, drop pipeline.rs's bytes-roundtrip
The recovery sweep's `_id` resolution was doing an awkward two-step:
open a `SuperfileReader`, immediately extract `parquet_bytes()`, throw
the reader away, then re-build a `ParquetRecordBatchReaderBuilder`
from those bytes to scan the column. Two parquet-footer parses per
superfile probed.

Move the scan onto the reader itself:

  - `SuperfileReader::id_lookup(target) -> Result<Option<u32>>`
    encapsulates the parquet open + Decimal128 sequential scan, using
    the reader's own `id_column` (set from the file's KV metadata at
    open time) so callers can't pass a mismatched column name.
  - `resolve_target_id_in_manifest` now keeps the `SuperfileReader`
    around and calls `reader.id_lookup(target)` directly; the
    storage-fallback branch opens a reader over the freshly-fetched
    bytes once instead of returning bare bytes for the caller to
    re-open.

This is API cleanup, not a perf win — `id_lookup` still re-builds the
parquet reader internally on each call. Caching parquet metadata on
`SuperfileReader` so `id_lookup` (and future scan-shaped methods)
skips the footer reparse is a separate follow-up.

Net -67 + scan helper inlined into the reader. 952/952 tests passing.
